improve error message for PHP < 5.4
When the user uses a PHP version smaller than 5.4, they are not able
to use the Symfony Installer. Instead, they are presented a nice error
message telling them to use the `composer create-project` command
instead. However, if that command is executed without a version
number, the user will get the latest stable release of the Symfony
Standard Edition. Especially when the user passed a custom version
number to the Symfony Installer, this might not be the desired
version. Therefore, it is reasonable to pass the required version
number to Composer.