Fix bug #symfony/12019
To fix this bug I have made an algorithm that sets the progress bar's width according to the columns of a terminal window.

The code uses external programs, "tput" for Linux/Unix systems and "mode" for Windows, to get the total columns of the window. In case of a terminal too wide, the code sets the width of the progress bar to the original value (60 chars). In case of a terminal too narrow, the code ignores the progress bar and displays only the file size and the percentage of the download. 

The bug was posted here: https://github.com/symfony/symfony/issues/13019